+++ src/Windows/Avalonia.Direct2D1/Media/GeometryImpl.cs
@@ -82,7 +82,7 @@ namespace Avalonia.Direct2D1.Media
             return false;
         }
 
-        public bool TryGetSegment(float startDistance, float stopDistance, bool startOnBeginFigure, out IGeometryImpl segmentGeometry)
+        public bool TryGetSegment(double startDistance, double stopDistance, bool startOnBeginFigure, out IGeometryImpl segmentGeometry)
         {
             // Direct2D doesnt have this too sadly.
             Logger.TryGet(LogEventLevel.Warning, LogArea.Visual)?.Log(this, "TryGetSegment is not available in Direct2D.");
