+++ tests/Avalonia.Visuals.UnitTests/VisualTree/MockRenderInterface.cs
@@ -163,7 +163,7 @@ namespace Avalonia.Visuals.UnitTests.VisualTree
                 throw new NotImplementedException();
             }
 
-            public bool TryGetSegment(float startDistance, float stopDistance, bool startOnBeginFigure, out IGeometryImpl segmentGeometry)
+            public bool TryGetSegment(double startDistance, double stopDistance, bool startOnBeginFigure, out IGeometryImpl segmentGeometry)
             {
                 throw new NotImplementedException();
             }
